refactor: move component HomepageFeatures to modules folder
